+++ packages/analytics-js-common/src/constants/integrations/config_to_integration_names.js
@@ -79,6 +79,7 @@ const configToIntNames = {
   SPOTIFYPIXEL: 'SpotifyPixel',
   COMMANDBAR: 'CommandBar',
   NINETAILED: 'Ninetailed',
+  XPIXEL: 'XPixel',
 };
 
 export { configToIntNames };