+++ packages/analytics-js-plugins/src/errorReporting/constants.ts
@@ -22,6 +22,7 @@ const REQUEST_TIMEOUT_MS = 10 * 1000; // 10 seconds
 const NOTIFIER_NAME = 'RudderStack JavaScript SDK Error Notifier';
 const SDK_GITHUB_URL = 'https://github.com/rudderlabs/rudder-sdk-js';
 const SOURCE_NAME = 'js';
+const ERROR_REPORTING_PLUGIN = 'ErrorReportingPlugin';
 
 export {
   SDK_FILE_NAME_PREFIXES,
@@ -31,4 +32,5 @@ export {
   NOTIFIER_NAME,
   SDK_GITHUB_URL,
   SOURCE_NAME,
+  ERROR_REPORTING_PLUGIN,
 };