+++ packages/analytics-js-cookies/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.3.5](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-cookies@0.3.4...@rudderstack/analytics-js-cookies@0.3.5) (2024-08-02)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.8.1`
 ## [0.3.4](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-cookies@0.3.3...@rudderstack/analytics-js-cookies@0.3.4) (2024-07-24)
 
 ### Dependency Updates