+++ packages/analytics-js-integrations/CHANGELOG.md
@@ -2,6 +2,28 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.7.0](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-integrations@3.6.0...@rudderstack/analytics-js-integrations@3.7.0) (2024-08-02)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.8.1`
+
+### Features
+
+* criteo add support for multiple hash methods ([#1812](https://github.com/rudderlabs/rudder-sdk-js/issues/1812)) ([23a0e37](https://github.com/rudderlabs/rudder-sdk-js/commit/23a0e37d274647cb3e2505694522080d7f5b98f3))
+
+
+### Bug Fixes
+
+* npm sanity suites ([#1810](https://github.com/rudderlabs/rudder-sdk-js/issues/1810)) ([22e43da](https://github.com/rudderlabs/rudder-sdk-js/commit/22e43da01f750a5cb23a2fce50de3744c54a197e))
+
+## [3.6.0](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-integrations@3.5.0...@rudderstack/analytics-js-integrations@3.6.0) (2024-08-01)
+
+
+### Features
+
+* **integrations:** feature for server side container url support ga4 ([cfbea6c](https://github.com/rudderlabs/rudder-sdk-js/commit/cfbea6c615e3ef63e06c0098ccdb19c70572c22a))
+
 ## [3.5.0](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-integrations@3.4.5...@rudderstack/analytics-js-integrations@3.5.0) (2024-07-24)
 
 ### Dependency Updates