+++ packages/analytics-js-plugins/CHANGELOG.md
@@ -2,6 +2,24 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.4.8](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-plugins@3.4.7...@rudderstack/analytics-js-plugins@3.4.8) (2024-08-16)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js` updated to version `3.7.7`
+## [3.4.7](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-plugins@3.4.6...@rudderstack/analytics-js-plugins@3.4.7) (2024-08-02)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.8.1`
+* `@rudderstack/analytics-js` updated to version `3.7.6`
+* `@rudderstack/analytics-js-cookies` updated to version `0.3.5`
+
+### Bug Fixes
+
+* error filtering of non-errors ([#1811](https://github.com/rudderlabs/rudder-sdk-js/issues/1811)) ([7b83e16](https://github.com/rudderlabs/rudder-sdk-js/commit/7b83e1661b1e0ce0b6b5ae45d3a2e08db97ddcb3))
+* npm sanity suites ([#1810](https://github.com/rudderlabs/rudder-sdk-js/issues/1810)) ([22e43da](https://github.com/rudderlabs/rudder-sdk-js/commit/22e43da01f750a5cb23a2fce50de3744c54a197e))
+
 ## [3.4.6](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-plugins@3.4.5...@rudderstack/analytics-js-plugins@3.4.6) (2024-07-24)
 
 ### Dependency Updates