+++ packages/analytics-js-service-worker/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.1.8](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-service-worker@3.1.7...@rudderstack/analytics-js-service-worker@3.1.8) (2024-08-16)
+
+
+### Bug Fixes
+
+* upgrade axios to a safe version ([#1827](https://github.com/rudderlabs/rudder-sdk-js/issues/1827)) ([bafe698](https://github.com/rudderlabs/rudder-sdk-js/commit/bafe6983c8e4db9e8055a96b8a7ab50c738b056d))
+
+## [3.1.7](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-service-worker@3.1.6...@rudderstack/analytics-js-service-worker@3.1.7) (2024-08-02)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.8.1`
 ## [3.1.6](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-service-worker@3.1.5...@rudderstack/analytics-js-service-worker@3.1.6) (2024-07-24)
 
 ### Dependency Updates